fix: extracted types to types folder
