+++ src/client/types/CourseRealisation.ts
@@ -1,0 +1,8 @@
+import { Locales } from './Locales'
+
+export type CourseRealisation = {
+  id: string
+  name: Locales
+  startDate: string
+  endDate: string
+}
