+++ src/client/types/DateRange.ts
@@ -1,0 +1,1 @@
+export type DateRange = { start: Date; end: Date }
