+++ src/client/types/FeedbackTarget.ts
@@ -1,0 +1,7 @@
+import { CourseRealisation } from './CourseRealisation'
+
+export type FeedbackTarget = {
+  id: number
+  courseRealisationId: string
+  courseRealisation: CourseRealisation
+}
