+++ src/client/types/FeedbackTargets.ts
@@ -1,0 +1,3 @@
+import { FeedbackTarget } from './FeedbackTarget'
+
+export type FeedbackTargets = FeedbackTarget[]
