+++ src/client/types/Locales.ts
@@ -1,0 +1,1 @@
+export type Locales = { fi: string; en: string; sv: string }
