+++ src/client/types/OrganisationCode.ts
@@ -1,0 +1,1 @@
+export type OrganisationCode = string
