+++ src/client/types/index.ts
@@ -1,0 +1,9 @@
+import { Locales } from './Locales'
+import { InitialValues } from './InitialValues'
+import { OrganisationCode } from './OrganisationCode'
+import { CourseRealisation } from './CourseRealisation'
+import { FeedbackTargets } from './FeedbackTargets'
+import { FeedbackTarget } from './FeedbackTarget'
+import { User } from './User'
+
+export type { Locales, InitialValues, OrganisationCode, CourseRealisation, FeedbackTargets, FeedbackTarget, User }
